Read description
Developed favorite/"unfavorite" projects feature. 
Also removed files not needed

+++ app/Http/Controllers/ProfileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Favorite;
 use App\Project;
 use App\Developer;
 use App\Follow;
@@ -173,9 +174,18 @@ class ProfileController extends Controller
             ]);
     }
 
-
-    public function follow(Request $request, $id_user)
+    /**
+     * Follows/Unfollows a user
+     *
+     * @param Request $request
+     * @param $id_user
+     * @return false|string
+     */
+    public function follow($id_user)
     {
+        if(empty(User::find($id_user)))
+            return json_encode(0);
+
         if(Follow::where([['id_follower', Auth::user()->getAuthIdentifier()],['id_followee', $id_user]])->exists()) {
             Follow::where([['id_follower', Auth::user()->getAuthIdentifier()],['id_followee', $id_user]])->delete();
         }
@@ -191,6 +201,25 @@ class ProfileController extends Controller
         return json_encode($id_user);
     }
 
+    public function favorite($id_project)
+    {
+        if(empty(Project::find($id_project)))
+            return json_encode(0);
+
+        if(Favorite::where([['id_user', Auth::user()->getAuthIdentifier()], ['id_project', $id_project]])->exists()) {
+            Favorite::where([['id_user', Auth::user()->getAuthIdentifier()], ['id_project', $id_project]])->delete();
+        }
+        else {
+            $favorite = new Favorite();
+
+            $favorite->id_user = Auth::user()->getAuthIdentifier();
+            $favorite->id_project = $id_project;
+
+            $favorite->save();
+        }
+
+        return json_encode($id_project);
+    }
 
     /**
      * Show the form for editing the specified resource.
